requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==0.28.0
pymongo==4.5.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.1.0
scikit-learn==1.3.0
flask-cors==4.0.0
gunicorn==21.2.0
